backend/requirements.txt
# Django Core
Django>=6.0.1
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg[binary]>=3.1.0

# Storage
django-storages>=1.14.0
boto3>=1.34.0

# Environment Management
python-dotenv>=1.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# API Documentation
drf-spectacular>=0.27.0
drf-spectacular-sidecar>=2023.12.1

# Caching & Background Jobs
django-redis>=5.4.0
celery>=5.4.0
redis>=5.0.0

# Development & Code Quality
black>=24.0.0
flake8>=7.0.0
isort>=5.13.0
pre-commit>=3.6.0
